+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This change log is manually updated at the moment.
 
 ## Unreleased
 
+## v1.5.1 (May 31st, 2022)
+
+- fix copy/paste in Firefox
+
 ## v1.5.0 (May 19th, 2022)
 
 - Wallet Connect